Introduced Application class
- also added some fresh icons
