+++ src/IconInfo.h
